Completate tutte le funzioni lato Server. Test eseguiti in modo molto blando. Sistemati alcuni bug delle strutture dati.

+++ lib/file_reader.h
@@ -8,6 +8,6 @@ FILE* fileReader(char* path);
 FILE* file_open(char* filename, char* mode);
 char* file_readline(FILE* file);
 void* file_readAll(FILE* file);
-int file_getsize(FILE* file);
+size_t file_getsize(FILE* file);
 void file_close(FILE* file);
 #endif //FILEREADER_FILE_READER_H
